removed unused stuff + moved db into schema file + broke todo tool into file
