+++ src/actions/delete-app.ts
@@ -2,7 +2,7 @@
 
 import { getUser } from "@/auth/stack-auth";
 import { appsTable, appUsers } from "@/db/schema";
-import { db } from "@/lib/db";
+import { db } from "@/db/schema";
 import { and, eq } from "drizzle-orm";
 import { freestyle } from "@/lib/freestyle";
