--- src/lib/db.ts
@@ -1,3 +1,0 @@
-import { drizzle } from "drizzle-orm/node-postgres";
-
-export const db = drizzle(process.env.DATABASE_URL!);
