--- src/lib/fs.ts
@@ -1,4 +1,0 @@
-"use client";
-import FS from "@isomorphic-git/lightning-fs";
-
-export const fs = new FS("adorable-fs");
