--- src/mastra/index.ts
@@ -1,3 +1,0 @@
-// import { Mastra } from '@mastra/core';
-
-// export const mastra = new Mastra()
